requirements.txt
Django==1.7.4
Markdown==2.5.2
boto==2.36.0
dj-database-url==0.3.0
django-appconf==0.6
django-filter==0.9.2
django-geoposition==0.2.2
django-storages==1.1.8
djangorestframework==3.0.4
psycopg2==2.5
six==1.9.0
wsgiref==0.1.2
